Utility methods for working with read preferences
convertType() is necessary because the driver returns numeric types in getReadPreference() instead of the string constants accepted by setReadPreference(). This is likely a bug in the early 1.3.x versions.

convertTagSets is necessary because the format returned by getReadPreference() differs from that accepted by setReadPreference(). This likely isn't a bug, but it's something to handle when we preserve tag sets in the setSlaveOkay() Database and Collection methods.