Don't set cursor flags unnecessarily and cast arguments
If slaveOkay defaults to false in Doctrine's class and is set each time the cursor is initialized, there is no way to inherit from MongoDB or MongoCollection ancestors. Instead, initialize these fields with empty values and only set them on the cursor if a non-null value is available.

Casting input arguments explicitly will ensure that any method call is sufficient to remove the null default.